+++ src/heap.c
@@ -16,7 +16,7 @@ int cmpBurstHeap(const Process *a, const Process *b)
     return a->remaining_burst - b->remaining_burst;
 }
 
-ProcessHeap* createHeap(int capacity, int heap_type)
+ProcessHeap* createHeap(size_t capacity, int heap_type)
 {
     ProcessHeap *ph = malloc(sizeof(ProcessHeap));
     if (ph == NULL)
@@ -53,18 +53,18 @@ void freeHeap(ProcessHeap *ph)
     free(ph);
 }
 
-void swap(Process **array, int a, int b)
+void swap(Process **array, size_t a, size_t b)
 {
     Process *temp = array[a];
     array[a] = array[b];
     array[b] = temp;
 }
 
-void minHeapify(ProcessHeap *ph, int index)
+void minHeapify(ProcessHeap *ph, size_t index)
 {
-    int min = index;
-    int left = index * 2 + 1;
-    int right = index * 2 + 2;
+    size_t min = index;
+    size_t left = index * 2 + 1;
+    size_t right = index * 2 + 2;
 
     if (left < ph->size && ph->cmp(ph->heap[left], ph->heap[min]) < 0)
         min = left;
@@ -87,7 +87,7 @@ void insertHeap(ProcessHeap *ph, Process *p)
     ph->heap[ph->size] = p;
     ph->size++;
 
-    int idx = ph->size - 1;
+    size_t idx = ph->size - 1;
     while (idx != 0 && ph->cmp(ph->heap[idx], ph->heap[(idx - 1) / 2]) < 0) {
         swap(ph->heap, idx, (idx - 1) / 2);
         idx = (idx - 1) / 2;
